Updated auth.js to get valid token(Issue #25)
Getting token with `user.xa` , `ya` and in my case `za` is a mess, these properties tend to change. This way the tokens are not renewed here, in fact those properties simply are reassigned to false, then a new property is assigned with the refreshed token, + this is also not a documented way to get token. 

Refactored `formatUser` to get valid token using `getIdToken()`  as mentioned in https://firebase.google.com/docs/reference/js/firebase.User#getidtoken